profiler/requirements.txt
torch
numpy
zeus-ml
matplotlib
